+++ engine/dt_send_eng.cpp
@@ -112,7 +112,7 @@ public:
 		// Remember where the game code pointed us for this datatable's data so 
 		m_pProxies[ pNode->GetRecursiveProxyIndex() ] = pStructBase;
 
-		for ( int iChild=0; iChild < pNode->GetNumChildren(); iChild++ )
+		for ( intp iChild=0; iChild < pNode->GetNumChildren(); iChild++ )
 		{
 			CSendNode *pCurChild = pNode->GetChild( iChild );
